pyproject.toml
[project]
name = "foreman"
version = "0.1.0"
description = "Event-driven backend for managing image-generation requests for AI models"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.2",
    "uvicorn[standard]>=0.40.0",
    "pydantic>=2.12.5",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-sdk>=1.39.1",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
    "opentelemetry-exporter-otlp>=1.39.1",
    "python-json-logger>=4.0.0",
    "asyncpg>=0.31.0",
    "psycopg[binary]>=3.2.10",
    "python-dotenv>=1.2.2",
    "boto3>=1.41.0",
    "google-genai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-asyncio>=1.3.0",
    "httpx>=0.27.2",
    "ruff>=0.15.7",
    "alembic>=1.18.4",
    "psycopg2-binary>=2.9.11",
    "testcontainers>=4.14.1",
    "sqlparse>=0.5.0",
    "moto>=5.0.0",
]

[build-system]
requires = ["setuptools>=80.10.2"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["foreman*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
testpaths = ["tests"]